refactor(API): Extract shared offset-cursor pagination helper
Lifts the duplicated cursor-decode-to-offset/limit logic out of the
tags, role-mapping-rules, and workflow-history public API list
handlers into resolveOffsetPagination(). getWorkflows keeps its own
inline variant since it has different fallback semantics.
