+++ packages/cli/src/public-api/v1/controllers/tags.public.controller.ts
@@ -12,8 +12,10 @@ import {
 } from '@n8n/decorators';
 import type { Response } from 'express';
 
-import { BadRequestError } from '@/errors/response-errors/bad-request.error';
-import { decodeCursor, encodeNextCursor } from '@/public-api/v1/shared/services/pagination.service';
+import {
+	encodeNextCursor,
+	resolveOffsetPagination,
+} from '@/public-api/v1/shared/services/pagination.service';
 import { TagService } from '@/services/tag.service';
 
 @PublicApiController('/tags')
@@ -31,22 +33,7 @@ export class TagsPublicController {
 		_res: Response,
 		@Query query: ListTagsQueryDto,
 	): Promise<TagListPublicDto> {
-		let offset = 0;
-		let { limit } = query;
-
-		if (query.cursor) {
-			try {
-				const decoded = decodeCursor(query.cursor);
-				if (!('offset' in decoded)) {
-					throw new BadRequestError('An invalid cursor was provided');
-				}
-				offset = decoded.offset;
-				limit = decoded.limit;
-			} catch (error) {
-				if (error instanceof BadRequestError) throw error;
-				throw new BadRequestError('An invalid cursor was provided');
-			}
-		}
+		const { offset, limit } = resolveOffsetPagination(query);
 
 		const { data, count } = await this.tagService.getPaginated({ offset, limit });
