+++ packages/cli/src/public-api/v1/shared/services/pagination.service.ts
@@ -1,5 +1,7 @@
 import { jsonParse } from 'n8n-workflow';
 
+import { BadRequestError } from '@/errors/response-errors/bad-request.error';
+
 import type {
 	CursorPagination,
 	OffsetPagination,
@@ -10,6 +12,35 @@ import type {
 export const decodeCursor = (cursor: string): PaginationOffsetDecoded | PaginationCursorDecoded => {
 	return jsonParse(Buffer.from(cursor, 'base64').toString());
 };
+
+/**
+ * Resolves the offset and limit to query with for a list endpoint either from defaults
+ * or from a provided cursor.
+ */
+export function resolveOffsetPagination(query: {
+	cursor?: string;
+	limit: number;
+	offset?: number;
+}): { offset: number; limit: number } {
+	let { limit } = query;
+	let offset = query.offset ?? 0;
+
+	if (query.cursor) {
+		try {
+			const decoded = decodeCursor(query.cursor);
+			if (!('offset' in decoded)) {
+				throw new BadRequestError('An invalid cursor was provided');
+			}
+			offset = decoded.offset;
+			limit = decoded.limit;
+		} catch (error) {
+			if (error instanceof BadRequestError) throw error;
+			throw new BadRequestError('An invalid cursor was provided');
+		}
+	}
+
+	return { offset, limit };
+}
 
 const encodeOffSetPagination = (pagination: OffsetPagination): string | null => {
 	if (pagination.numberOfTotalRecords > pagination.offset + pagination.limit) {
